refactor(core): 迁移工具辅助函数到 memory/turn_result.rs 并添加注释
- 将 turn 记忆记录、媒体解析、工具产物提取等 ~20 个函数集中到 memory/turn_result.rs
- 为每个函数添加文档注释说明作用
- react/engine.rs 中的引用路径更新为 crate::memory::turn_result::
- core/mod.rs 从 1445 行降至 838 行（累计从 4375 行降 81%）
