+++ crates/tiangong-core/src/memory/turn_result.rs
@@ -3,15 +3,19 @@
 //! 负责从会话消息中提取工具调用、媒体产物、文件产物等 TurnArtifact，
 //! 构建 TurnResult 供 Memory Actor 写入长期记忆。
 //!
-//! TODO: Phase 2 后续步骤中将 core/mod.rs 中的旧实现迁移到此处。
-
-#![allow(dead_code)]
+//! 同时提供工具结果的媒体解析、图片归档等辅助能力，
+//! 供 ReactEngine 在工具执行完毕后调用。
 
 use std::collections::HashSet;
 
 use crate::session::{Message, MessageRole};
 
-/// 构建 TurnResult，记录一轮对话的关键产物和工具调用。
+// ── Turn 记忆记录函数 ──
+
+/// 构建一轮对话的 TurnResult，提取工具调用、媒体/文件产物并生成摘要。
+///
+/// 从 `session.messages[turn_start_idx..]` 中提取当前轮次的所有信息，
+/// 生成唯一的 `turn_id`，组装为 `TurnResult` 供 Memory Actor 写入长期记忆。
 pub(crate) fn build_memory_turn_result(
     session: &crate::session::Session,
     turn_start_idx: usize,
@@ -33,6 +37,10 @@ pub(crate) fn build_memory_turn_result(
     }
 }
 
+/// 从消息列表中提取所有去重的工具调用名称。
+///
+/// 仅扫描 `Tool` 角色的消息，通过 `parse_tool_calls_line` 和 `parse_tool_trace_name`
+/// 两种格式提取工具名，自动过滤 `recall_memory`，返回去重后的名称列表。
 fn extract_turn_tool_calls(messages: &[Message]) -> Vec<String> {
     let mut seen = HashSet::new();
     let mut names = Vec::new();
@@ -55,6 +63,11 @@ fn extract_turn_tool_calls(messages: &[Message]) -> Vec<String> {
     names
 }
 
+/// 从消息列表中提取所有媒体、文件、工具结果类型的 TurnArtifact。
+///
+/// 扫描 `Assistant` 消息中的 media 字段和 `Tool` 消息中的工具追踪内容，
+/// 提取图片/视频 URL、写入文件路径、工具结果摘要等产物，
+/// 通过 URL/path 去重，最多返回 12 个产物。
 fn extract_turn_artifacts(messages: &[Message]) -> Vec<tiangong_memory::TurnArtifact> {
     let mut artifacts = Vec::new();
     let mut seen = HashSet::new();
@@ -131,6 +144,10 @@ fn extract_turn_artifacts(messages: &[Message]) -> Vec<tiangong_memory::TurnArti
     artifacts.into_iter().take(12).collect()
 }
 
+/// 构建轮次摘要文本。
+///
+/// 优先取最后一条非空 Assistant 消息的压缩内容作为摘要；
+/// 如果没有 Assistant 消息，则将所有产物的 URL/path/summary 拼接为摘要。
 fn build_turn_memory_summary(
     messages: &[Message],
     artifacts: &[tiangong_memory::TurnArtifact],
@@ -160,6 +177,12 @@ fn build_turn_memory_summary(
         .join("\n")
 }
 
+// ── 解析辅助函数 ──
+
+/// 解析 `tool_calls:` 前缀行，提取逗号分隔的工具调用名称列表。
+///
+/// 例如 `tool_calls: read_file, write_file` 返回 `["read_file", "write_file"]`。
+/// 如果不存在该前缀行，返回空 Vec。
 fn parse_tool_calls_line(content: &str) -> Vec<String> {
     content
         .lines()
@@ -174,6 +197,9 @@ fn parse_tool_calls_line(content: &str) -> Vec<String> {
         .unwrap_or_default()
 }
 
+/// 从工具追踪消息中提取工具名称。
+///
+/// 匹配格式为 `工具执行 [tool_name]` 的首行，返回方括号内的工具名。
 pub(crate) fn parse_tool_trace_name(content: &str) -> Option<String> {
     let first_line = content.lines().next()?.trim();
     let rest = first_line.strip_prefix("工具执行 [")?;
@@ -181,6 +207,10 @@ pub(crate) fn parse_tool_trace_name(content: &str) -> Option<String> {
     Some(rest[..end].to_string()).filter(|name| !name.is_empty())
 }
 
+/// 解析 `summary:` 前缀行，提取单行摘要文本。
+///
+/// 返回第一个匹配的 summary 行内容（去除前后空白），
+/// 如果不存在或为空则返回 None。
 fn parse_summary_line(content: &str) -> Option<String> {
     content.lines().find_map(|line| {
         line.trim()
@@ -191,6 +221,31 @@ fn parse_summary_line(content: &str) -> Option<String> {
     })
 }
 
+/// 将文本压缩到指定字符数以内。
+///
+/// 先去除空行并合并为单段文本，如果字符数超过 `max_chars` 则截断并追加 `...`。
+/// 用于生成记忆摘要等场景，避免过长文本占用存储空间。
+pub(crate) fn compact_single_memory_text(text: &str, max_chars: usize) -> String {
+    let normalized = text
+        .lines()
+        .map(str::trim)
+        .filter(|line| !line.is_empty())
+        .collect::<Vec<_>>()
+        .join("\n");
+    if normalized.chars().count() <= max_chars {
+        return normalized;
+    }
+    let mut clipped = normalized.chars().take(max_chars).collect::<String>();
+    clipped.push_str("...");
+    clipped
+}
+
+// ── 媒体产物解析函数 ──
+
+/// 从工具追踪消息中解析媒体类型的 TurnArtifact。
+///
+/// 逐行扫描内容，提取 Markdown 图片语法 `![title](url)` 和视频 URL 行，
+/// 将其转换为 `TurnArtifactKind::Media` 类型的产物条目。
 fn parse_media_artifacts_from_tool_trace(
     content: &str,
     tool_name: Option<&str>,
@@ -233,6 +288,170 @@ fn parse_media_artifacts_from_tool_trace(
     artifacts
 }
 
+/// 从工具结果中解析 MediaAsset 列表（图片或视频）。
+///
+/// 根据 `tool_name` 判断输出类型：
+/// - 如果输出可能包含生成的图片，则解析图片资源；
+/// - 如果工具名为 `generate_video`，则解析视频资源；
+/// - 其他情况返回空列表。
+pub(crate) fn parse_media_assets_from_tool_result(
+    tool_name: &str,
+    stdout: &str,
+    summary: &str,
+) -> Vec<tiangong_types::MediaAsset> {
+    if output_may_contain_generated_images(tool_name, stdout) {
+        return parse_image_assets(stdout);
+    }
+    if tool_name == "generate_video" {
+        parse_video_assets(stdout, summary)
+    } else {
+        Vec::new()
+    }
+}
+
+/// 将工具结果中的图片 URL 归档到本地存储。
+///
+/// 仅当工具执行成功且输出可能包含生成的图片时才执行归档，
+/// 归档后替换 `result.stdout` 中的远程 URL 为本地路径。
+pub(crate) fn localize_tool_result_images(tool_name: &str, result: &mut crate::tool::ToolResult) {
+    if !result.ok || !output_may_contain_generated_images(tool_name, &result.stdout) {
+        return;
+    }
+    result.stdout = archive_image_markdown_output(&result.stdout);
+}
+
+/// 判断工具输出是否可能包含生成的图片。
+///
+/// 满足以下任一条件即返回 true：
+/// - 工具名为 `generate_image`；
+/// - 输出内容是纯图片 Markdown 格式；
+/// - 工具名包含 "image" 且输出包含 `](`（Markdown 图片链接特征）。
+fn output_may_contain_generated_images(tool_name: &str, output: &str) -> bool {
+    tool_name == "generate_image"
+        || looks_like_pure_image_markdown(output)
+        || (tool_name.to_ascii_lowercase().contains("image") && output.contains("]("))
+}
+
+/// 判断输出是否是纯图片 Markdown 格式。
+///
+/// 即所有非空行都是 `![alt](url)` 格式时返回 true。
+fn looks_like_pure_image_markdown(output: &str) -> bool {
+    let trimmed = output.trim();
+    !trimmed.is_empty()
+        && trimmed.lines().all(|line| {
+            let line = line.trim();
+            line.is_empty()
+                || (line.starts_with("![") && line.contains("](") && line.ends_with(')'))
+        })
+}
+
+/// 将图片 Markdown 输出中的 URL 归档到本地存储。
+///
+/// 逐行扫描输出文本，对每行 Markdown 图片调用 `archive_image_reference` 归档，
+/// 归档成功则替换为本地路径，失败则保留原始 URL 并打印警告日志。
+fn archive_image_markdown_output(output: &str) -> String {
+    output
+        .lines()
+        .map(|line| {
+            let Some((alt, url)) = parse_markdown_image_line(line.trim()) else {
+                return line.to_string();
+            };
+            match crate::media_archive::archive_image_reference(url, None) {
+                Ok(archived) => format!("![{alt}]({})", archived.path()),
+                Err(err) => {
+                    tracing::warn!(url = %url, error = %err, "图片归档到本地失败，保留原始 URL");
+                    line.to_string()
+                }
+            }
+        })
+        .collect::<Vec<_>>()
+        .join("\n")
+}
+
+/// 解析单行 Markdown 图片语法，提取 alt 文本和 URL。
+///
+/// 匹配格式 `![alt](url)`，返回 `(alt, url)` 元组。
+/// 如果行不符合图片 Markdown 格式或 URL 为空，返回 None。
+fn parse_markdown_image_line(line: &str) -> Option<(&str, &str)> {
+    if !line.starts_with("![") || !line.ends_with(')') {
+        return None;
+    }
+    let close_alt = line.find("](")?;
+    let alt = line[2..close_alt].trim();
+    let url = line[close_alt + 2..line.len() - 1].trim();
+    (!url.is_empty()).then_some((alt, url))
+}
+
+/// 从输出文本中解析所有图片类型的 MediaAsset。
+///
+/// 逐行扫描 Markdown 图片格式 `![title](url)`，
+/// 同时识别 data URI 中的 MIME 类型（如 `data:image/png;...`）。
+/// 每个匹配行生成一个 `MediaKind::Image` 类型的资源条目。
+fn parse_image_assets(output: &str) -> Vec<tiangong_types::MediaAsset> {
+    output
+        .lines()
+        .filter_map(|line| {
+            let line = line.trim();
+            if !line.starts_with("![") || !line.ends_with(')') {
+                return None;
+            }
+            let close_alt = line.find("](")?;
+            let title = line[2..close_alt].trim();
+            let url = line[close_alt + 2..line.len() - 1].trim();
+            if url.is_empty() {
+                return None;
+            }
+            let mime_type = url
+                .strip_prefix("data:")
+                .and_then(|raw| raw.split(';').next())
+                .filter(|mime| mime.starts_with("image/"))
+                .map(str::to_string);
+            Some(tiangong_types::MediaAsset {
+                kind: tiangong_types::MediaKind::Image,
+                url: url.to_string(),
+                mime_type,
+                title: (!title.is_empty()).then(|| title.to_string()),
+                capability: Some("image_generation".to_string()),
+            })
+        })
+        .collect()
+}
+
+/// 从输出文本中解析所有视频类型的 MediaAsset。
+///
+/// 逐行扫描视频 URL 行，将每个 URL 转换为 `MediaKind::Video` 类型的资源条目，
+/// MIME 类型默认为 `video/mp4`。
+fn parse_video_assets(output: &str, summary: &str) -> Vec<tiangong_types::MediaAsset> {
+    output
+        .lines()
+        .filter_map(|line| parse_video_url_line(line.trim()))
+        .map(|url| tiangong_types::MediaAsset {
+            kind: tiangong_types::MediaKind::Video,
+            url,
+            mime_type: Some("video/mp4".to_string()),
+            title: Some(summary.to_string()).filter(|item| !item.trim().is_empty()),
+            capability: Some("video_generation".to_string()),
+        })
+        .collect()
+}
+
+/// 解析视频 URL 行。
+///
+/// 匹配 `Video URL: https://...` 或 `video_url: https://...` 前缀，
+/// 提取第一个以 `http://` 或 `https://` 开头的 URL。
+fn parse_video_url_line(line: &str) -> Option<String> {
+    let raw = line
+        .strip_prefix("Video URL:")
+        .or_else(|| line.strip_prefix("video_url:"))
+        .map(str::trim)?;
+    let url = raw.split_whitespace().next().unwrap_or(raw);
+    (url.starts_with("http://") || url.starts_with("https://")).then(|| url.to_string())
+}
+
+/// 解析工具追踪消息中的写入文件路径。
+///
+/// 从 `命令:` 前缀行中提取 `path=...` 参数值，
+/// 用于记录文件写入操作的目标路径。
 fn parse_written_path(content: &str) -> Option<String> {
     let command = content.lines().find_map(|line| {
         line.trim()
@@ -245,6 +464,11 @@ fn parse_written_path(content: &str) -> Option<String> {
     Some(rest[..end].trim().to_string()).filter(|path| !path.is_empty())
 }
 
+/// 判断工具结果是否应记录为 TurnArtifact。
+///
+/// 只读类工具（`read_file`、`list_dir`、`tree_dir`、`search_code`、
+/// `recall_memory`、`get_skill_detail`）的结果不需要记录，
+/// 其他工具的结果均应记录。
 fn should_record_tool_result(tool_name: &str) -> bool {
     !matches!(
         tool_name,
@@ -255,28 +479,4 @@ fn should_record_tool_result(tool_name: &str) -> bool {
             | "recall_memory"
             | "get_skill_detail"
     )
-}
-
-pub(crate) fn compact_single_memory_text(text: &str, max_chars: usize) -> String {
-    let normalized = text
-        .lines()
-        .map(str::trim)
-        .filter(|line| !line.is_empty())
-        .collect::<Vec<_>>()
-        .join("\n");
-    if normalized.chars().count() <= max_chars {
-        return normalized;
-    }
-    let mut clipped = normalized.chars().take(max_chars).collect::<String>();
-    clipped.push_str("...");
-    clipped
-}
-
-fn parse_video_url_line(line: &str) -> Option<String> {
-    let raw = line
-        .strip_prefix("Video URL:")
-        .or_else(|| line.strip_prefix("video_url:"))
-        .map(str::trim)?;
-    let url = raw.split_whitespace().next().unwrap_or(raw);
-    (url.starts_with("http://") || url.starts_with("https://")).then(|| url.to_string())
 }
